Demote inline-blame down the activation order
= Since this plugin's functionality is non-critical, activation should
be delayed as much as possible, achieved by the following and heavy use
of `requestIdleCallback`:

Load rest of module only when needed (ie. not when loading empty
workspace).
Only activate on TextEditor, without blocking.
Lazily attach to current editor.

= Use CSS variables to not add and remove a whole <style> tag.

= Remove line number dependence.

Sometimes the style line doesn't match the current line in the
editor, causing blame to not be shown when it's still correct.

= use Set.

Should benchmark but probably not needed unless we expect a user
to open 1000+ editors.